requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
autoscraper
pollinations.ai
pollinations
packaging
pydantic-settings
python-dotenv
duckduckgo-search
ddgs
urllib3
cloudscraper 
